+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 3.1.2
 - phishing 0.20.4
 - hw-ledger, keyring, util, util-crypto 11.1.2
 - extension-dapp 0.45.3